add fork handler test
