+++ network/p2p/p2p_validation_test.go
@@ -14,15 +14,13 @@ import (
 	"time"
 
 	"github.com/aquasecurity/table"
-	"github.com/cornelk/hashmap"
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
-	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/sourcegraph/conc/pool"
-	spectypes "github.com/ssvlabs/ssv-spec/types"
 	"github.com/stretchr/testify/require"
 
+	spectypes "github.com/ssvlabs/ssv-spec/types"
+	spectestingutils "github.com/ssvlabs/ssv-spec/types/testingutils"
 	"github.com/ssvlabs/ssv/message/validation"
-	"github.com/ssvlabs/ssv/protocol/v2/ssv/queue"
 )
 
 // TestP2pNetwork_MessageValidation tests p2pNetwork would score peers according
@@ -49,65 +47,19 @@ func TestP2pNetwork_MessageValidation(t *testing.T) {
 		cryptorand.Read(validator[:])
 		validators[i] = hex.EncodeToString(validator[:])
 	}
-
+	var mtx sync.Mutex
 	// Create a MessageValidator to accept/reject/ignore messages according to their role type.
 	const (
 		acceptedRole = spectypes.RoleProposer
 		ignoredRole  = spectypes.RoleAggregator
 		rejectedRole = spectypes.RoleSyncCommitteeContribution
 	)
-	messageValidators := make([]*MockMessageValidator, nodeCount)
-	var mtx sync.Mutex
-	for i := 0; i < nodeCount; i++ {
-		i := i
-		messageValidators[i] = &MockMessageValidator{
-			Accepted: make([]int, nodeCount),
-			Ignored:  make([]int, nodeCount),
-			Rejected: make([]int, nodeCount),
-		}
-		messageValidators[i].ValidateFunc = func(ctx context.Context, p peer.ID, pmsg *pubsub.Message) pubsub.ValidationResult {
-			peer := vNet.NodeByPeerID(p)
-			signedSSVMsg := &spectypes.SignedSSVMessage{}
-			require.NoError(t, signedSSVMsg.Decode(pmsg.GetData()))
-
-			decodedMsg, err := queue.DecodeSignedSSVMessage(signedSSVMsg)
-			require.NoError(t, err)
-			pmsg.ValidatorData = decodedMsg
-			mtx.Lock()
-			// Validation according to role.
-			var validation pubsub.ValidationResult
-			switch signedSSVMsg.SSVMessage.MsgID.GetRoleType() {
-			case acceptedRole:
-				messageValidators[i].Accepted[peer.Index]++
-				messageValidators[i].TotalAccepted++
-				validation = pubsub.ValidationAccept
-			case ignoredRole:
-				messageValidators[i].Ignored[peer.Index]++
-				messageValidators[i].TotalIgnored++
-				validation = pubsub.ValidationIgnore
-			case rejectedRole:
-				messageValidators[i].Rejected[peer.Index]++
-				messageValidators[i].TotalRejected++
-				validation = pubsub.ValidationReject
-			default:
-				panic("unsupported role")
-			}
-			mtx.Unlock()
-
-			// Always accept messages from self to make libp2p propagate them,
-			// while still counting them by their role.
-			if p == vNet.Nodes[i].Network.Host().ID() {
-				return pubsub.ValidationAccept
-			}
-
-			return validation
-		}
-	}
-
+	messageValidators := CreateMsgValidators(&mtx, nodeCount, vNet)
 	// Create a VirtualNet with 4 nodes.
+	ks := spectestingutils.Testing4SharesSet()
 	vNet = CreateVirtualNet(t, ctx, 4, validators, func(nodeIndex int) validation.MessageValidator {
 		return messageValidators[nodeIndex]
-	})
+	}, ks)
 	defer func() {
 		require.NoError(t, vNet.Close())
 	}()
@@ -251,117 +203,4 @@ func TestP2pNetwork_MessageValidation(t *testing.T) {
 		}
 	}
 	defer fmt.Println()
-}
-
-type MockMessageValidator struct {
-	Accepted      []int
-	Ignored       []int
-	Rejected      []int
-	TotalAccepted int
-	TotalIgnored  int
-	TotalRejected int
-
-	ValidateFunc func(ctx context.Context, p peer.ID, pmsg *pubsub.Message) pubsub.ValidationResult
-}
-
-func (v *MockMessageValidator) ValidatorForTopic(topic string) func(ctx context.Context, p peer.ID, pmsg *pubsub.Message) pubsub.ValidationResult {
-	return v.Validate
-}
-
-func (v *MockMessageValidator) Validate(ctx context.Context, p peer.ID, pmsg *pubsub.Message) pubsub.ValidationResult {
-	return v.ValidateFunc(ctx, p, pmsg)
-}
-
-type NodeIndex int
-
-type VirtualNode struct {
-	Index      NodeIndex
-	Network    *p2pNetwork
-	PeerScores *hashmap.Map[NodeIndex, *pubsub.PeerScoreSnapshot]
-}
-
-func (n *VirtualNode) Broadcast(msgID spectypes.MessageID, msg *spectypes.SignedSSVMessage) error {
-	return n.Network.Broadcast(msgID, msg)
-}
-
-// VirtualNet is a utility to create & interact with a virtual network of nodes.
-type VirtualNet struct {
-	Nodes []*VirtualNode
-}
-
-func CreateVirtualNet(
-	t *testing.T,
-	ctx context.Context,
-	nodes int,
-	validatorPubKeys []string,
-	messageValidatorProvider func(int) validation.MessageValidator,
-) *VirtualNet {
-	var doneSetup atomic.Bool
-	vn := &VirtualNet{}
-	ln, routers, err := createNetworkAndSubscribe(t, ctx, LocalNetOptions{
-		Nodes:                    nodes,
-		MinConnected:             nodes - 1,
-		UseDiscv5:                false,
-		TotalValidators:          1000,
-		ActiveValidators:         800,
-		MyValidators:             300,
-		MessageValidatorProvider: messageValidatorProvider,
-		PeerScoreInspector: func(selfPeer peer.ID, peerMap map[peer.ID]*pubsub.PeerScoreSnapshot) {
-			if !doneSetup.Load() {
-				return
-			}
-			node := vn.NodeByPeerID(selfPeer)
-			if node == nil {
-				t.Fatalf("self peer not found (%s)", selfPeer)
-			}
-
-			node.PeerScores.Range(func(index NodeIndex, snapshot *pubsub.PeerScoreSnapshot) bool {
-				node.PeerScores.Del(index)
-				return true
-			})
-			for peerID, peerScore := range peerMap {
-				peerNode := vn.NodeByPeerID(peerID)
-				if peerNode == nil {
-					t.Fatalf("peer not found (%s)", peerID)
-				}
-				node.PeerScores.Set(peerNode.Index, peerScore)
-			}
-
-		},
-		PeerScoreInspectorInterval: time.Millisecond * 5,
-	}, validatorPubKeys...)
-
-	require.NoError(t, err)
-	require.NotNil(t, routers)
-	require.NotNil(t, ln)
-
-	for i, node := range ln.Nodes {
-		vn.Nodes = append(vn.Nodes, &VirtualNode{
-			Index:      NodeIndex(i),
-			Network:    node.(*p2pNetwork),
-			PeerScores: hashmap.New[NodeIndex, *pubsub.PeerScoreSnapshot](), //{}make(map[NodeIndex]*pubsub.PeerScoreSnapshot),
-		})
-	}
-	doneSetup.Store(true)
-
-	return vn
-}
-
-func (vn *VirtualNet) NodeByPeerID(peerID peer.ID) *VirtualNode {
-	for _, node := range vn.Nodes {
-		if node.Network.Host().ID() == peerID {
-			return node
-		}
-	}
-	return nil
-}
-
-func (vn *VirtualNet) Close() error {
-	for _, node := range vn.Nodes {
-		err := node.Network.Close()
-		if err != nil {
-			return err
-		}
-	}
-	return nil
 }
